Remove Python locks from the destroy method.
Deadlock issue was reported by Forest Bond.

NOTES:

These locks were undisputedly bogus as the point when destroy is called
is out of the scope where Python locking is set up.

Removing them and doing nothing more (as it's done now) is also OK if we assume
that the module is used for writing a standalone program where thread management
is completely handled by the fuse library -- as in this case the lib will ensure
that there will be only one thread when the destroy method is called.

However, this way:
 - We rely on an implementation detail of the lib.
 - If the Python fuse module is intended to be used in a "library fashion",
   ie. as a part of a more complex program where threads are created apart
   from the fuselib ones, then the locks might be needed for destroy.

Anyway, I couldn't really extend the scope of Python locking in a way
that embraces destroy's invocation and actually works, so I stick with
simply removing those locks for now.