use bytes objects for read and write data
This changes the data type for the buffer of the read and write syscalls
from string to bytes. On Python 2, this has no effect. On Python 3, it
is a breaking change, but fixes a serious usibility bug that limits file
data to valid UTF-8 data. With these changes, files can contain arbitrary
data.