Define PY_SSIZE_T_CLEAN for Python 3.10
On Python 3.10, `PY_SSIZE_T_CLEAN` must be defined in order to use '#'
formats.  `ioctl_func` needs a slight change to comply with the new
regime.

(Several functions pass `size_t` here, which may be a little dodgy, but
the main point of this change seems to have been to transition from
`int` to `Py_ssize_t`, and at least `size_t` and `Py_ssize_t` have the
same size.)

Closes #41.