Patch 1/3 - Replace `Optparse` with `ArgumentParser`
This is the first of a 3-patch series of commits to address libfuse/python-fuse#24

This commit introduces tests for the current `Optparse` based implementation
to ensure that the next commit doesn't break current usage compatibility.

The next commit would introduce the `ArgumentParser` based implementation. The
final commit will attempt to do some minor clean up to make the code a bit
more idomatic / pythonic.